+++ tests/testthat/test-shapes.R
@@ -62,3 +62,14 @@ test_that("cleveland_shape_pal works with overlap = FALSE", {
 test_that("scale_shape_cleveland works", {
   expect_s3_class(scale_shape_cleveland(), "ScaleDiscrete")
 })
+
+test_that("warn_unicode_pch warns only for unicode pch on non-UTF-8 locales", {
+  testthat::local_mocked_bindings(l10n_info = function() list(`UTF-8` = FALSE), .package = "base")
+  expect_warning(warn_unicode_pch(c(-9675, -9679)), "Unicode")
+  expect_no_warning(warn_unicode_pch(c(1, 2, 3)))
+})
+
+test_that("warn_unicode_pch is silent on UTF-8 locales", {
+  testthat::local_mocked_bindings(l10n_info = function() list(`UTF-8` = TRUE), .package = "base")
+  expect_no_warning(warn_unicode_pch(c(-9675, -9679)))
+})